+++ framework/src/main/groovy/org/moqui/impl/service/ServiceJsonRpcDispatcher.groovy
@@ -60,6 +60,7 @@ public class ServiceJsonRpcDispatcher {
                 } else {
                     jsonRespList.add(callSingle(null, callSingleObj, null))
                 }
+                eci.getMessage().clearAll()
             }
             eci.getWeb().sendJsonResponse(jsonRespList)
         } else {